done with file attachements to courses and components questions and a…

+++ resources/views/notes/notes.blade.php
@@ -11,7 +11,7 @@
       <h3><a href="/notes/view_note_details/{{$note->id}}">{{$note->title}}</a></h3>
       <p>{!! nl2br(e($note->description)) !!}</p>
       @if((!empty($role))&&$role==1)
-          <a href="/admin/delete_note/{{$note->id}}">Delete</a>
+          <a onclick="return confirm('Are you sure want to delete this note?');" href="/admin/delete_note/{{$note->id}}">Delete</a>
       @endif
       <!-- /browse/notes/view_note/{{$note->id}} -->
     </li>
